+++ react/lib/FullscreenIcon.tsx
@@ -1,28 +1,27 @@
-import React, { useEffect, useImperativeHandle, useRef } from 'react';
-import 'jb-grid';
-import type { JBFullscreenIconAttributes } from './module-declaration.js';
+import React, { useEffect, useImperativeHandle, useRef } from "react";
+import "jb-icons/expand";
+import type { JBFullscreenIconAttributes } from "./module-declaration.js";
+import type { JBIconExpandWebComponent } from "jb-icons/expand";
 
 type JBFullscreenIconState = "enter" | "exit";
-type JBFullscreenIconWebComponent = HTMLElement & {
-  state: JBFullscreenIconState
-}
+type JBFullscreenIconWebComponent = JBIconExpandWebComponent;
 
 export type FullscreenIconProps = Omit<JBFullscreenIconAttributes, "ref" | "state"> & {
-  state?: JBFullscreenIconState
-}
+  state?: JBFullscreenIconState;
+};
 
 export const JBFullscreenIcon = React.forwardRef<JBFullscreenIconWebComponent | null, FullscreenIconProps>((props, ref) => {
   const { state, ...otherProps } = props;
   const element = useRef<JBFullscreenIconWebComponent | null>(null);
-  useImperativeHandle(ref, () => element.current as JBFullscreenIconWebComponent, [element]);
+  useImperativeHandle(ref, () => element.current as JBFullscreenIconWebComponent, []);
 
   useEffect(() => {
     if (element.current) {
-      element.current.state = state ?? "enter";
+      element.current.isExpanded = state === "exit";
     }
   }, [state]);
 
-  return <jb-fullscreen-icon ref={element} {...otherProps}></jb-fullscreen-icon>;
+  return <jb-icon-expand ref={element} {...otherProps}></jb-icon-expand>;
 });
 
 JBFullscreenIcon.displayName = "JBFullscreenIcon";