+++ react/lib/jb-grid/i18n.ts
@@ -4,6 +4,8 @@ export type JBGridDictionary = {
   closeFilters: string;
   openFilters: string;
   refresh: string;
+  enterFullscreen: string;
+  exitFullscreen: string;
 };
 
 export const gridDictionary = new JBDictionary<JBGridDictionary>({
@@ -11,10 +13,14 @@ export const gridDictionary = new JBDictionary<JBGridDictionary>({
     closeFilters: "بستن فیلترها",
     openFilters: "باز کردن فیلترها",
     refresh: "تازه‌سازی داده‌ها",
+    enterFullscreen: "ورود به حالت تمام‌صفحه",
+    exitFullscreen: "خروج از حالت تمام‌صفحه",
   },
   en: {
     closeFilters: "Close filters",
     openFilters: "Open filters",
     refresh: "Refresh data",
+    enterFullscreen: "Enter fullscreen",
+    exitFullscreen: "Exit fullscreen",
   },
 });