+++ web-component/lib/column-header/render.ts
@@ -5,7 +5,7 @@ export function renderHTML(): string {
         <slot name="title"></slot>
       </span>
       <span class="sort-icon-wrapper" part="sort-icon" aria-hidden="true">
-        <jb-icon-arrow-tailed direction="up"></jb-icon-arrow-tailed>
+        <jb-icon-arrow-tailed direction="up" long></jb-icon-arrow-tailed>
       </span>
     </div>
       `;